+++ sdk/resourcemanager/containerinstance/armcontainerinstance/models_serde.go
@@ -18,7 +18,7 @@ import (
 
 // MarshalJSON implements the json.Marshaller interface for type AzureFileVolume.
 func (a AzureFileVolume) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "readOnly", a.ReadOnly)
 	populate(objectMap, "shareName", a.ShareName)
 	populate(objectMap, "storageAccountKey", a.StorageAccountKey)
@@ -57,7 +57,7 @@ func (a *AzureFileVolume) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CachedImages.
 func (c CachedImages) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "image", c.Image)
 	populate(objectMap, "osType", c.OSType)
 	return json.Marshal(objectMap)
@@ -88,7 +88,7 @@ func (c *CachedImages) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CachedImagesListResult.
 func (c CachedImagesListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", c.NextLink)
 	populate(objectMap, "value", c.Value)
 	return json.Marshal(objectMap)
@@ -119,7 +119,7 @@ func (c *CachedImagesListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Capabilities.
 func (c Capabilities) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "capabilities", c.Capabilities)
 	populate(objectMap, "gpu", c.Gpu)
 	populate(objectMap, "ipAddressType", c.IPAddressType)
@@ -166,7 +166,7 @@ func (c *Capabilities) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CapabilitiesCapabilities.
 func (c CapabilitiesCapabilities) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "maxCpu", c.MaxCPU)
 	populate(objectMap, "maxGpuCount", c.MaxGpuCount)
 	populate(objectMap, "maxMemoryInGB", c.MaxMemoryInGB)
@@ -201,7 +201,7 @@ func (c *CapabilitiesCapabilities) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type CapabilitiesListResult.
 func (c CapabilitiesListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", c.NextLink)
 	populate(objectMap, "value", c.Value)
 	return json.Marshal(objectMap)
@@ -232,7 +232,7 @@ func (c *CapabilitiesListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ConfidentialComputeProperties.
 func (c ConfidentialComputeProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "ccePolicy", c.CcePolicy)
 	return json.Marshal(objectMap)
 }
@@ -259,7 +259,7 @@ func (c *ConfidentialComputeProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Container.
 func (c Container) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "name", c.Name)
 	populate(objectMap, "properties", c.Properties)
 	return json.Marshal(objectMap)
@@ -290,7 +290,7 @@ func (c *Container) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerAttachResponse.
 func (c ContainerAttachResponse) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "password", c.Password)
 	populate(objectMap, "webSocketUri", c.WebSocketURI)
 	return json.Marshal(objectMap)
@@ -321,7 +321,7 @@ func (c *ContainerAttachResponse) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerExec.
 func (c ContainerExec) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "command", c.Command)
 	return json.Marshal(objectMap)
 }
@@ -348,7 +348,7 @@ func (c *ContainerExec) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerExecRequest.
 func (c ContainerExecRequest) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "command", c.Command)
 	populate(objectMap, "terminalSize", c.TerminalSize)
 	return json.Marshal(objectMap)
@@ -379,7 +379,7 @@ func (c *ContainerExecRequest) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerExecRequestTerminalSize.
 func (c ContainerExecRequestTerminalSize) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "cols", c.Cols)
 	populate(objectMap, "rows", c.Rows)
 	return json.Marshal(objectMap)
@@ -410,7 +410,7 @@ func (c *ContainerExecRequestTerminalSize) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerExecResponse.
 func (c ContainerExecResponse) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "password", c.Password)
 	populate(objectMap, "webSocketUri", c.WebSocketURI)
 	return json.Marshal(objectMap)
@@ -441,7 +441,7 @@ func (c *ContainerExecResponse) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroup.
 func (c ContainerGroup) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", c.ID)
 	populate(objectMap, "identity", c.Identity)
 	populate(objectMap, "location", c.Location)
@@ -496,7 +496,7 @@ func (c *ContainerGroup) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupDiagnostics.
 func (c ContainerGroupDiagnostics) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "logAnalytics", c.LogAnalytics)
 	return json.Marshal(objectMap)
 }
@@ -523,7 +523,7 @@ func (c *ContainerGroupDiagnostics) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupIdentity.
 func (c ContainerGroupIdentity) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "principalId", c.PrincipalID)
 	populate(objectMap, "tenantId", c.TenantID)
 	populate(objectMap, "type", c.Type)
@@ -562,7 +562,7 @@ func (c *ContainerGroupIdentity) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupListResult.
 func (c ContainerGroupListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", c.NextLink)
 	populate(objectMap, "value", c.Value)
 	return json.Marshal(objectMap)
@@ -593,7 +593,7 @@ func (c *ContainerGroupListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupProperties.
 func (c ContainerGroupProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "identity", c.Identity)
 	populate(objectMap, "properties", c.Properties)
 	return json.Marshal(objectMap)
@@ -624,7 +624,7 @@ func (c *ContainerGroupProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupPropertiesInstanceView.
 func (c ContainerGroupPropertiesInstanceView) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "events", c.Events)
 	populate(objectMap, "state", c.State)
 	return json.Marshal(objectMap)
@@ -655,7 +655,7 @@ func (c *ContainerGroupPropertiesInstanceView) UnmarshalJSON(data []byte) error 
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupPropertiesProperties.
 func (c ContainerGroupPropertiesProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "confidentialComputeProperties", c.ConfidentialComputeProperties)
 	populate(objectMap, "containers", c.Containers)
 	populate(objectMap, "dnsConfig", c.DNSConfig)
@@ -746,7 +746,7 @@ func (c *ContainerGroupPropertiesProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerGroupSubnetID.
 func (c ContainerGroupSubnetID) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", c.ID)
 	populate(objectMap, "name", c.Name)
 	return json.Marshal(objectMap)
@@ -777,7 +777,7 @@ func (c *ContainerGroupSubnetID) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerHTTPGet.
 func (c ContainerHTTPGet) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "httpHeaders", c.HTTPHeaders)
 	populate(objectMap, "path", c.Path)
 	populate(objectMap, "port", c.Port)
@@ -816,7 +816,7 @@ func (c *ContainerHTTPGet) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerPort.
 func (c ContainerPort) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "port", c.Port)
 	populate(objectMap, "protocol", c.Protocol)
 	return json.Marshal(objectMap)
@@ -847,7 +847,7 @@ func (c *ContainerPort) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerProbe.
 func (c ContainerProbe) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "exec", c.Exec)
 	populate(objectMap, "failureThreshold", c.FailureThreshold)
 	populate(objectMap, "httpGet", c.HTTPGet)
@@ -898,7 +898,7 @@ func (c *ContainerProbe) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerProperties.
 func (c ContainerProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "command", c.Command)
 	populate(objectMap, "environmentVariables", c.EnvironmentVariables)
 	populate(objectMap, "image", c.Image)
@@ -957,7 +957,7 @@ func (c *ContainerProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerPropertiesInstanceView.
 func (c ContainerPropertiesInstanceView) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "currentState", c.CurrentState)
 	populate(objectMap, "events", c.Events)
 	populate(objectMap, "previousState", c.PreviousState)
@@ -996,7 +996,7 @@ func (c *ContainerPropertiesInstanceView) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ContainerState.
 func (c ContainerState) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "detailStatus", c.DetailStatus)
 	populate(objectMap, "exitCode", c.ExitCode)
 	populateTimeRFC3339(objectMap, "finishTime", c.FinishTime)
@@ -1039,7 +1039,7 @@ func (c *ContainerState) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type DNSConfiguration.
 func (d DNSConfiguration) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nameServers", d.NameServers)
 	populate(objectMap, "options", d.Options)
 	populate(objectMap, "searchDomains", d.SearchDomains)
@@ -1074,7 +1074,7 @@ func (d *DNSConfiguration) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type DeploymentExtensionSpec.
 func (d DeploymentExtensionSpec) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "name", d.Name)
 	populate(objectMap, "properties", d.Properties)
 	return json.Marshal(objectMap)
@@ -1105,7 +1105,7 @@ func (d *DeploymentExtensionSpec) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type DeploymentExtensionSpecProperties.
 func (d DeploymentExtensionSpecProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "extensionType", d.ExtensionType)
 	populate(objectMap, "protectedSettings", &d.ProtectedSettings)
 	populate(objectMap, "settings", &d.Settings)
@@ -1144,7 +1144,7 @@ func (d *DeploymentExtensionSpecProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type EncryptionProperties.
 func (e EncryptionProperties) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "identity", e.Identity)
 	populate(objectMap, "keyName", e.KeyName)
 	populate(objectMap, "keyVersion", e.KeyVersion)
@@ -1183,7 +1183,7 @@ func (e *EncryptionProperties) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type EnvironmentVariable.
 func (e EnvironmentVariable) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "name", e.Name)
 	populate(objectMap, "secureValue", e.SecureValue)
 	populate(objectMap, "value", e.Value)
@@ -1218,7 +1218,7 @@ func (e *EnvironmentVariable) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Event.
 func (e Event) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "count", e.Count)
 	populateTimeRFC3339(objectMap, "firstTimestamp", e.FirstTimestamp)
 	populateTimeRFC3339(objectMap, "lastTimestamp", e.LastTimestamp)
@@ -1265,7 +1265,7 @@ func (e *Event) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type GitRepoVolume.
 func (g GitRepoVolume) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "directory", g.Directory)
 	populate(objectMap, "repository", g.Repository)
 	populate(objectMap, "revision", g.Revision)
@@ -1300,7 +1300,7 @@ func (g *GitRepoVolume) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type GpuResource.
 func (g GpuResource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "count", g.Count)
 	populate(objectMap, "sku", g.SKU)
 	return json.Marshal(objectMap)
@@ -1331,7 +1331,7 @@ func (g *GpuResource) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type HTTPHeader.
 func (h HTTPHeader) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "name", h.Name)
 	populate(objectMap, "value", h.Value)
 	return json.Marshal(objectMap)
@@ -1362,7 +1362,7 @@ func (h *HTTPHeader) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type IPAddress.
 func (i IPAddress) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "autoGeneratedDomainNameLabelScope", i.AutoGeneratedDomainNameLabelScope)
 	populate(objectMap, "dnsNameLabel", i.DNSNameLabel)
 	populate(objectMap, "fqdn", i.Fqdn)
@@ -1409,7 +1409,7 @@ func (i *IPAddress) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ImageRegistryCredential.
 func (i ImageRegistryCredential) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "identity", i.Identity)
 	populate(objectMap, "identityUrl", i.IdentityURL)
 	populate(objectMap, "password", i.Password)
@@ -1452,7 +1452,7 @@ func (i *ImageRegistryCredential) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type InitContainerDefinition.
 func (i InitContainerDefinition) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "name", i.Name)
 	populate(objectMap, "properties", i.Properties)
 	return json.Marshal(objectMap)
@@ -1483,7 +1483,7 @@ func (i *InitContainerDefinition) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type InitContainerPropertiesDefinition.
 func (i InitContainerPropertiesDefinition) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "command", i.Command)
 	populate(objectMap, "environmentVariables", i.EnvironmentVariables)
 	populate(objectMap, "image", i.Image)
@@ -1526,7 +1526,7 @@ func (i *InitContainerPropertiesDefinition) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type InitContainerPropertiesDefinitionInstanceView.
 func (i InitContainerPropertiesDefinitionInstanceView) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "currentState", i.CurrentState)
 	populate(objectMap, "events", i.Events)
 	populate(objectMap, "previousState", i.PreviousState)
@@ -1565,7 +1565,7 @@ func (i *InitContainerPropertiesDefinitionInstanceView) UnmarshalJSON(data []byt
 
 // MarshalJSON implements the json.Marshaller interface for type LogAnalytics.
 func (l LogAnalytics) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "logType", l.LogType)
 	populate(objectMap, "metadata", l.Metadata)
 	populate(objectMap, "workspaceId", l.WorkspaceID)
@@ -1608,7 +1608,7 @@ func (l *LogAnalytics) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Logs.
 func (l Logs) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "content", l.Content)
 	return json.Marshal(objectMap)
 }
@@ -1635,7 +1635,7 @@ func (l *Logs) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Operation.
 func (o Operation) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "display", o.Display)
 	populate(objectMap, "name", o.Name)
 	populate(objectMap, "origin", o.Origin)
@@ -1674,7 +1674,7 @@ func (o *Operation) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type OperationDisplay.
 func (o OperationDisplay) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "description", o.Description)
 	populate(objectMap, "operation", o.Operation)
 	populate(objectMap, "provider", o.Provider)
@@ -1713,7 +1713,7 @@ func (o *OperationDisplay) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type OperationListResult.
 func (o OperationListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "nextLink", o.NextLink)
 	populate(objectMap, "value", o.Value)
 	return json.Marshal(objectMap)
@@ -1744,7 +1744,7 @@ func (o *OperationListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Port.
 func (p Port) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "port", p.Port)
 	populate(objectMap, "protocol", p.Protocol)
 	return json.Marshal(objectMap)
@@ -1775,7 +1775,7 @@ func (p *Port) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Resource.
 func (r Resource) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "id", r.ID)
 	populate(objectMap, "location", r.Location)
 	populate(objectMap, "name", r.Name)
@@ -1822,7 +1822,7 @@ func (r *Resource) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceLimits.
 func (r ResourceLimits) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "cpu", r.CPU)
 	populate(objectMap, "gpu", r.Gpu)
 	populate(objectMap, "memoryInGB", r.MemoryInGB)
@@ -1857,7 +1857,7 @@ func (r *ResourceLimits) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceRequests.
 func (r ResourceRequests) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "cpu", r.CPU)
 	populate(objectMap, "gpu", r.Gpu)
 	populate(objectMap, "memoryInGB", r.MemoryInGB)
@@ -1892,7 +1892,7 @@ func (r *ResourceRequests) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type ResourceRequirements.
 func (r ResourceRequirements) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "limits", r.Limits)
 	populate(objectMap, "requests", r.Requests)
 	return json.Marshal(objectMap)
@@ -1923,7 +1923,7 @@ func (r *ResourceRequirements) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Usage.
 func (u Usage) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "currentValue", u.CurrentValue)
 	populate(objectMap, "id", u.ID)
 	populate(objectMap, "limit", u.Limit)
@@ -1966,7 +1966,7 @@ func (u *Usage) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type UsageListResult.
 func (u UsageListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "value", u.Value)
 	return json.Marshal(objectMap)
 }
@@ -1993,7 +1993,7 @@ func (u *UsageListResult) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type UsageName.
 func (u UsageName) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "localizedValue", u.LocalizedValue)
 	populate(objectMap, "value", u.Value)
 	return json.Marshal(objectMap)
@@ -2024,7 +2024,7 @@ func (u *UsageName) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type UserAssignedIdentities.
 func (u UserAssignedIdentities) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "clientId", u.ClientID)
 	populate(objectMap, "principalId", u.PrincipalID)
 	return json.Marshal(objectMap)
@@ -2055,7 +2055,7 @@ func (u *UserAssignedIdentities) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type Volume.
 func (v Volume) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "azureFile", v.AzureFile)
 	populate(objectMap, "emptyDir", &v.EmptyDir)
 	populate(objectMap, "gitRepo", v.GitRepo)
@@ -2098,7 +2098,7 @@ func (v *Volume) UnmarshalJSON(data []byte) error {
 
 // MarshalJSON implements the json.Marshaller interface for type VolumeMount.
 func (v VolumeMount) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
+	objectMap := make(map[string]any)
 	populate(objectMap, "mountPath", v.MountPath)
 	populate(objectMap, "name", v.Name)
 	populate(objectMap, "readOnly", v.ReadOnly)
@@ -2131,7 +2131,7 @@ func (v *VolumeMount) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-func populate(m map[string]interface{}, k string, v interface{}) {
+func populate(m map[string]any, k string, v any) {
 	if v == nil {
 		return
 	} else if azcore.IsNullValue(v) {
@@ -2141,7 +2141,7 @@ func populate(m map[string]interface{}, k string, v interface{}) {
 	}
 }
 
-func unpopulate(data json.RawMessage, fn string, v interface{}) error {
+func unpopulate(data json.RawMessage, fn string, v any) error {
 	if data == nil {
 		return nil
 	}