+++ sdk/resourcemanager/customerlockbox/armcustomerlockbox/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 0.6.0 (2023-03-28)
+### Breaking Changes
+
+- Struct `ErrorAdditionalInfo` has been removed
+
+### Features Added
+
+- New struct `ClientFactory` which is a client factory used to create any client in this module
+
+
 ## 0.5.0 (2022-05-18)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/customerlockbox/armcustomerlockbox` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 0.5.0, which contains breaking changes.